[operator] celoss finished without label_smoothing
