+++ src/flag_gems/ops/cross_entropy_loss.py
@@ -1,358 +1,372 @@
 import logging
-from enum import IntEnum
 
 import torch
 import triton
 import triton.language as tl
 
-from ..utils import libentry
-from .sum import sum, sum_dim
+from .sum import sum
 
 
-class Reduction(IntEnum):
-    NONE = 0
-    MEAN = 1
-    SUM = 2
-
-
-def heur_block_n(args):
-    return triton.next_power_of_2(args["N"])
-
-
-def heur_num_warps(args):
-    if args["N"] <= 1024:
-        return 4
-    elif args["N"] <= 2048:
-        return 8
-    else:
-        return 16
-
-
-@libentry()
 @triton.autotune(
     configs=[
-        triton.Config({"BLOCK_M": 1}, num_stages=4),
-        triton.Config({"BLOCK_M": 1}, num_stages=5),
-        triton.Config({"BLOCK_M": 2}, num_stages=4),
-        triton.Config({"BLOCK_M": 2}, num_stages=5),
-        triton.Config({"BLOCK_M": 4}, num_stages=4),
-        triton.Config({"BLOCK_M": 4}, num_stages=5),
-        triton.Config({"BLOCK_M": 8}, num_stages=4),
-        triton.Config({"BLOCK_M": 8}, num_stages=5),
+        triton.Config({"BLOCK_C": c, "BLOCK_D": d}, num_warps=4)
+        for c in [256, 512, 1024]
+        for d in [1, 4, 16]
     ],
-    key=[
-        "M",
-        "N",
-    ],
+    key=["C", "D"],
 )
-@triton.heuristics(
-    {
-        "BLOCK_N": heur_block_n,
-        "num_warps": heur_num_warps,
-    }
-)
-@triton.jit(do_not_specialize=["mean_num"])
-def log_softmax_and_mul_kernel(
-    output_ptr,
-    input_ptr,
-    target_ptr,
-    mean_num,
-    M,
+@triton.jit
+def celoss_indice_kernel(
+    inp_ptr,
+    tgt_ptr,
+    w_ptr,
+    out_ptr,
+    w_tgt_ptr,
+    ignore_index,
     N,
-    K,
-    BLOCK_M: tl.constexpr,
-    BLOCK_N: tl.constexpr,
+    C,
+    D,
+    BLOCK_C: tl.constexpr,
+    BLOCK_D: tl.constexpr,
 ):
-    pid_m = tl.program_id(0)
-    pid_k = tl.program_id(1)
-    m_offset = pid_m * BLOCK_M + tl.arange(0, BLOCK_M)
-    n_offset = tl.arange(0, BLOCK_N)
-    offset = m_offset[:, None] * N * K + n_offset[None, :] * K + pid_k
-    mask = m_offset[:, None] < M and n_offset[None, :] < N
-    input_ptrs = input_ptr + offset
-    inp = tl.load(input_ptrs, mask=mask, other=-float("inf"))
-    row_minus_max = inp - tl.max(inp, axis=1)[:, None]
-    numerator = tl.exp(row_minus_max)
-    denominator = tl.sum(numerator, axis=1)[:, None]
-    softmax_output = tl.log(numerator / denominator)
-    target = tl.load(target_ptr + offset, mask=mask, other=0.0)
-    out = softmax_output * target / (mean_num)
-    output_ptrs = output_ptr + offset
-    tl.store(output_ptrs, out, mask=mask)
+    pid_n = tl.program_id(0)
+    pid_d = tl.program_id(1)
+    offset_d = pid_d * BLOCK_D + tl.arange(0, BLOCK_D)
+
+    tgt_ptrs = tgt_ptr + pid_n * D + offset_d
+    tgt_mask = offset_d < D
+    tgt = tl.load(tgt_ptrs, mask=tgt_mask, other=0)
+
+    ignore_mask = not (tgt == ignore_index)
+
+    w_ptrs = w_ptr + tgt
+    w_tgt = tl.load(w_ptrs, mask=tgt_mask, other=0).to(tl.float32)
+    w_tgt_ptrs = w_tgt_ptr + pid_n * D + offset_d
+    tl.store(w_tgt_ptrs, w_tgt, mask=tgt_mask and ignore_mask)
+
+    tmp_max = tl.zeros([BLOCK_C, BLOCK_D], dtype=tl.float32)
+    tmp_sum = tl.zeros([BLOCK_C, BLOCK_D], dtype=tl.float32)
+
+    for off in range(0, C, BLOCK_C):
+        offset_c = off + tl.arange(0, BLOCK_C)
+        inp_ptrs = inp_ptr + pid_n * C * D + offset_c[:, None] * D + offset_d[None, :]
+        inp_mask = offset_c[:, None] < C and offset_d[None, :] < D
+        inp = tl.load(inp_ptrs, inp_mask, other=-float("inf")).to(tl.float32)
+        cur_max = tl.maximum(tmp_max, inp)
+        cur_exp = tl.exp(inp - cur_max)
+        tmp_sum = tmp_sum * tl.exp(tmp_max - cur_max) + cur_exp
+        tmp_max = cur_max
+    final_max = tl.max(tmp_max, axis=0)
+    tmp_sum = tmp_sum * tl.exp(tmp_max - final_max[None, :])
+    final_sum = tl.log(tl.sum(tmp_sum, axis=0))
+
+    inp_tgt_ptrs = inp_ptr + pid_n * C * D + tgt * D + offset_d
+    inp_tgt = tl.load(inp_tgt_ptrs, mask=tgt_mask, other=-float("inf")).to(tl.float32)
+
+    out = (final_sum + final_max - inp_tgt) * w_tgt
+    out_ptrs = out_ptr + pid_n * D + offset_d
+    tl.store(out_ptrs, out, mask=tgt_mask and ignore_mask)
 
 
-@libentry()
 @triton.autotune(
     configs=[
-        triton.Config({"BLOCK_M": 1}, num_stages=4),
-        triton.Config({"BLOCK_M": 1}, num_stages=5),
-        triton.Config({"BLOCK_M": 2}, num_stages=4),
-        triton.Config({"BLOCK_M": 2}, num_stages=5),
-        triton.Config({"BLOCK_M": 4}, num_stages=4),
-        triton.Config({"BLOCK_M": 4}, num_stages=5),
-        triton.Config({"BLOCK_M": 8}, num_stages=4),
-        triton.Config({"BLOCK_M": 8}, num_stages=5),
+        triton.Config({"BLOCK_C": c, "BLOCK_D": d}, num_warps=4)
+        for c in [256, 512, 1024]
+        for d in [1, 4, 16]
     ],
-    key=[
-        "M",
-        "N",
-    ],
+    key=["C", "D"],
 )
-@triton.heuristics(
-    {
-        "BLOCK_N": heur_block_n,
-        "num_warps": heur_num_warps,
-    }
-)
-@triton.jit(do_not_specialize=["mean_num"])
-def softmax_and_sub_kernel(
-    output_ptr,
-    input_ptr,
-    target_ptr,
-    out_grad,
-    mean_num,
-    M,
+@triton.jit
+def celoss_probability_kernel(
+    inp_ptr,
+    tgt_ptr,
+    w_ptr,
+    out_ptr,
     N,
-    K,
-    BLOCK_M: tl.constexpr,
-    BLOCK_N: tl.constexpr,
+    C,
+    D,
+    BLOCK_C: tl.constexpr,
+    BLOCK_D: tl.constexpr,
 ):
-    pid_m = tl.program_id(0)
-    pid_k = tl.program_id(1)
-    m_offset = pid_m * BLOCK_M + tl.arange(0, BLOCK_M)
-    n_offset = tl.arange(0, BLOCK_N)
-    offset = m_offset[:, None] * N * K + n_offset[None, :] * K + pid_k
-    mask = m_offset[:, None] < M and n_offset[None, :] < N
-    input_ptrs = input_ptr + offset
-    inp = tl.load(input_ptrs, mask=mask, other=-float("inf"))
-    row_minus_max = inp - tl.max(inp, axis=1)[:, None]
-    numerator = tl.exp(row_minus_max)
-    denominator = tl.sum(numerator, axis=1)[:, None]
-    # todo: reduce unnecessary calculations through mask operations to improve performance
-    softmax_output = numerator / denominator
-    target_ptrs = target_ptr + offset
-    target = tl.load(target_ptrs, mask=mask, other=0.0)
-    out_grad_ptr = out_grad + m_offset[:, None] * K + pid_k
-    out_grad_value = tl.load(out_grad_ptr)
-    # backward formula derivation for value of ingnore index
-    target_sum = tl.sum(target, axis=1)[:, None]
-    out = out_grad_value * (target_sum * softmax_output - target) / mean_num
+    pid_n = tl.program_id(0)
+    pid_d = tl.program_id(1)
+    offset_d = pid_d * BLOCK_D + tl.arange(0, BLOCK_D)
 
-    output_ptrs = output_ptr + offset
-    tl.store(output_ptrs, out, mask=mask)
+    tmp_max = tl.zeros([BLOCK_C, BLOCK_D], dtype=tl.float32)
+    tmp_sum = tl.zeros([BLOCK_C, BLOCK_D], dtype=tl.float32)
+
+    for off in range(0, C, BLOCK_C):
+        offset_c = off + tl.arange(0, BLOCK_C)
+        inp_ptrs = inp_ptr + pid_n * C * D + offset_c[:, None] * D + offset_d[None, :]
+        inp_mask = offset_c[:, None] < C and offset_d[None, :] < D
+        inp = tl.load(inp_ptrs, inp_mask, other=-float("inf")).to(tl.float32)
+        cur_max = tl.maximum(tmp_max, inp)
+        cur_exp = tl.exp(inp - cur_max)
+        tmp_sum = tmp_sum * tl.exp(tmp_max - cur_max) + cur_exp
+        tmp_max = cur_max
+    final_max = tl.max(tmp_max, axis=0)
+    tmp_sum = tmp_sum * tl.exp(tmp_max - final_max[None, :])
+    final_sum = tl.log(tl.sum(tmp_sum, axis=0))
+
+    _sum = tl.zeros([BLOCK_C, BLOCK_D], dtype=tl.float32)
+    for off in range(0, C, BLOCK_C):
+        offset_c = off + tl.arange(0, BLOCK_C)
+        inp_ptrs = inp_ptr + pid_n * C * D + offset_c[:, None] * D + offset_d[None, :]
+        tgt_ptrs = tgt_ptr + pid_n * C * D + offset_c[:, None] * D + offset_d[None, :]
+        mask = offset_c[:, None] < C and offset_d[None, :] < D
+        w_ptrs = w_ptr + offset_c
+        w_mask = offset_c < C
+        inp = tl.load(inp_ptrs, mask, other=0).to(tl.float32)
+        tgt = tl.load(tgt_ptrs, mask, other=1).to(tl.float32)
+        w = tl.load(w_ptrs, w_mask, other=0).to(tl.float32)[:, None]
+        log = final_sum[None, :] + final_max[None, :] - inp
+        _sum += w * log * tgt
+
+    out = tl.sum(_sum, axis=0)
+    out_ptrs = out_ptr + pid_n * D + offset_d
+    tl.store(out_ptrs, out, mask=offset_d < D)
 
 
-@libentry()
 @triton.autotune(
     configs=[
-        triton.Config({"BLOCK_M": 1}, num_stages=4),
-        triton.Config({"BLOCK_M": 1}, num_stages=5),
-        triton.Config({"BLOCK_M": 2}, num_stages=4),
-        triton.Config({"BLOCK_M": 2}, num_stages=5),
-        triton.Config({"BLOCK_M": 4}, num_stages=4),
-        triton.Config({"BLOCK_M": 4}, num_stages=5),
-        triton.Config({"BLOCK_M": 8}, num_stages=4),
-        triton.Config({"BLOCK_M": 8}, num_stages=5),
+        triton.Config({"BLOCK_C": c, "BLOCK_D": d}, num_warps=4)
+        for c in [256, 512, 1024]
+        for d in [1, 4, 16]
     ],
-    key=[
-        "M",
-        "N",
-    ],
+    key=["C", "D"],
 )
-@triton.heuristics(
-    {
-        "BLOCK_N": heur_block_n,
-        "num_warps": heur_num_warps,
-    }
-)
-@triton.jit(do_not_specialize=["mean_num"])
-def softmax_and_sub_reduce_kernel(
-    output_ptr,
-    input_ptr,
-    target_ptr,
-    out_grad,
+@triton.jit
+def celoss_indice_bwd(
+    out_grad_ptr,
+    inp_ptr,
+    tgt_ptr,
+    w_ptr,
+    inp_grad_ptr,
+    ignore_index,
     mean_num,
-    M,
     N,
-    K,
-    BLOCK_M: tl.constexpr,
-    BLOCK_N: tl.constexpr,
+    C,
+    D,
+    BLOCK_C: tl.constexpr,
+    BLOCK_D: tl.constexpr,
 ):
-    pid_m = tl.program_id(0)
-    pid_k = tl.program_id(1)
-    m_offset = pid_m * BLOCK_M + tl.arange(0, BLOCK_M)
-    n_offset = tl.arange(0, BLOCK_N)
-    offset = m_offset[:, None] * N * K + n_offset[None, :] * K + pid_k
-    mask = m_offset[:, None] < M and n_offset[None, :] < N
-    input_ptrs = input_ptr + offset
-    inp = tl.load(input_ptrs, mask=mask, other=-float("inf"))
-    row_minus_max = inp - tl.max(inp, axis=1)[:, None]
-    numerator = tl.exp(row_minus_max)
-    denominator = tl.sum(numerator, axis=1)[:, None]
-    # todo: reduce unnecessary calculations through mask operations to improve performance
-    softmax_output = numerator / denominator
-    target_ptrs = target_ptr + offset
-    target = tl.load(target_ptrs, mask=mask, other=0.0)
-    # backward formula derivation for value of ingnore index
-    target_sum = tl.sum(target, axis=1)[:, None]
-    out_grad_value = tl.load(out_grad)
-    out = out_grad_value * (target_sum * softmax_output - target) / mean_num
+    pid_n = tl.program_id(0)
+    pid_d = tl.program_id(1)
+    offset_d = pid_d * BLOCK_D + tl.arange(0, BLOCK_D)
 
-    output_ptrs = output_ptr + offset
-    tl.store(output_ptrs, out, mask=mask)
+    tgt_ptrs = tgt_ptr + pid_n * D + offset_d
+    tgt_mask = offset_d < D
+    tgt = tl.load(tgt_ptrs, mask=tgt_mask, other=0)
+    out_grad_ptrs = out_grad_ptr + pid_n * D + offset_d
+    out_grad = tl.load(out_grad_ptrs, mask=tgt_mask, other=0).to(tl.float32)[None, :]
+    w_ptrs = w_ptr + tgt
+    w_tgt = tl.load(w_ptrs, mask=tgt_mask, other=0).to(tl.float32)[None, :]
+
+    ignore_mask = (tgt != ignore_index)[None, :]
+
+    tmp_max = tl.zeros([BLOCK_C, BLOCK_D], dtype=tl.float32)
+    tmp_sum = tl.zeros([BLOCK_C, BLOCK_D], dtype=tl.float32)
+
+    for off in range(0, C, BLOCK_C):
+        offset_c = off + tl.arange(0, BLOCK_C)
+        inp_ptrs = inp_ptr + pid_n * C * D + offset_c[:, None] * D + offset_d[None, :]
+        inp_mask = offset_c[:, None] < C and offset_d[None, :] < D
+        inp = tl.load(inp_ptrs, inp_mask, other=-float("inf")).to(tl.float32)
+        cur_max = tl.maximum(tmp_max, inp)
+        cur_exp = tl.exp(inp - cur_max)
+        tmp_sum = tmp_sum * tl.exp(tmp_max - cur_max) + cur_exp
+        tmp_max = cur_max
+    final_max = tl.max(tmp_max, axis=0)
+    tmp_sum = tmp_sum * tl.exp(tmp_max - final_max[None, :])
+    final_sum = tl.sum(tmp_sum, axis=0)
+
+    for off in range(0, C, BLOCK_C):
+        offset_c = off + tl.arange(0, BLOCK_C)
+        inp_ptrs = inp_ptr + pid_n * C * D + offset_c[:, None] * D + offset_d[None, :]
+        inp_mask = offset_c[:, None] < C and offset_d[None, :] < D
+        inp = tl.load(inp_ptrs, inp_mask, other=-float("inf")).to(tl.float32)
+        minus_one = offset_c[:, None] == tgt[None, :]
+        inp_grad = (
+            (tl.exp(inp - final_max[None, :]) / final_sum[None, :] - minus_one)
+            * w_tgt
+            * out_grad
+            * mean_num
+        )
+        inp_grad_ptrs = (
+            inp_grad_ptr + pid_n * C * D + offset_c[:, None] * D + offset_d[None, :]
+        )
+        tl.store(inp_grad_ptrs, inp_grad, mask=inp_mask and ignore_mask)
+
+
+@triton.autotune(
+    configs=[
+        triton.Config({"BLOCK_C": c, "BLOCK_D": d}, num_warps=4)
+        for c in [256, 512, 1024]
+        for d in [1, 4, 16]
+    ],
+    key=["C", "D"],
+)
+@triton.jit
+def celoss_probability_bwd(
+    out_grad_ptr,
+    inp_ptr,
+    tgt_ptr,
+    w_ptr,
+    inp_grad_ptr,
+    mean_num,
+    N,
+    C,
+    D,
+    BLOCK_C: tl.constexpr,
+    BLOCK_D: tl.constexpr,
+):
+    pid_n = tl.program_id(0)
+    pid_d = tl.program_id(1)
+    offset_d = pid_d * BLOCK_D + tl.arange(0, BLOCK_D)
+
+    out_grad_ptrs = out_grad_ptr + pid_n * D + offset_d
+    out_grad = tl.load(out_grad_ptrs, mask=offset_d < D, other=0).to(tl.float32)[
+        None, :
+    ]
+
+    tmp_max = tl.zeros([BLOCK_C, BLOCK_D], dtype=tl.float32)
+    tmp_sum = tl.zeros([BLOCK_C, BLOCK_D], dtype=tl.float32)
+    w_tgt_sum = tl.zeros([BLOCK_C, BLOCK_D], dtype=tl.float32)
+
+    for off in range(0, C, BLOCK_C):
+        offset_c = off + tl.arange(0, BLOCK_C)
+        mask = offset_c[:, None] < C and offset_d[None, :] < D
+        inp_ptrs = inp_ptr + pid_n * C * D + offset_c[:, None] * D + offset_d[None, :]
+        inp = tl.load(inp_ptrs, mask, other=-float("inf")).to(tl.float32)
+
+        tgt_ptrs = tgt_ptr + pid_n * C * D + offset_c[:, None] * D + offset_d[None, :]
+        tgt = tl.load(tgt_ptrs, mask, other=0).to(tl.float32)
+
+        w_ptrs = w_ptr + offset_c
+        w_mask = offset_c < C
+        w = tl.load(w_ptrs, w_mask, other=0).to(tl.float32)[:, None]
+
+        w_tgt_sum += tgt * w
+
+        cur_max = tl.maximum(tmp_max, inp)
+        cur_exp = tl.exp(inp - cur_max)
+        tmp_sum = tmp_sum * tl.exp(tmp_max - cur_max) + cur_exp
+        tmp_max = cur_max
+    final_max = tl.max(tmp_max, axis=0)
+    tmp_sum = tmp_sum * tl.exp(tmp_max - final_max[None, :])
+    final_sum = tl.sum(tmp_sum, axis=0)[None, :]
+    w_tgt_sum = tl.sum(w_tgt_sum, axis=0)[None, :]
+
+    for off in range(0, C, BLOCK_C):
+        offset_c = off + tl.arange(0, BLOCK_C)
+        offset = pid_n * C * D + offset_c[:, None] * D + offset_d[None, :]
+        inp_ptrs = inp_ptr + offset
+        mask = offset_c[:, None] < C and offset_d[None, :] < D
+        inp = tl.load(inp_ptrs, mask, other=0).to(tl.float32)
+
+        tgt_ptrs = tgt_ptr + offset
+        tgt = tl.load(tgt_ptrs, mask, other=0).to(tl.float32)
+
+        w_ptrs = w_ptr + offset_c
+        w_mask = offset_c < C
+        w = tl.load(w_ptrs, w_mask, other=0).to(tl.float32)[:, None]
+
+        grad = w_tgt_sum / final_sum * tl.exp(inp - final_max) - w * tgt
+        inp_grad = grad * out_grad * mean_num
+
+        inp_grad_ptrs = inp_grad_ptr + offset
+        tl.store(inp_grad_ptrs, inp_grad, mask)
 
 
 class CrossEntropyLoss(torch.autograd.Function):
     @staticmethod
-    def forward(ctx, input, target, weight, reduction, ignore_index, label_smoothing):
+    def forward(ctx, inp, target, weight, reduction, ignore_index, label_smoothing):
         logging.debug("GEMS CrossEntropyLoss")
-        assert reduction in Reduction._value2member_map_, "Invalid reduction"
-        assert isinstance(input, torch.Tensor), "input is not a tensor"
+        # label_smoothing not supported
 
-        if input.ndim >= 2:
-            dim = 1
-        else:
-            dim = 0
+        shape = list(inp.shape)
+        dim = inp.ndim
+        N = 1 if dim == 1 else shape[0]
+        C = shape[0] if dim == 1 else shape[1]
+        D = inp.numel() // N // C
+        axis = 0 if dim == 1 else 1
+        del shape[axis]
 
-        if reduction != Reduction.MEAN.value:
-            mean_num = -1
-        else:
-            # get all ingore count of target
-            ignore_count = (target == ignore_index).to(torch.int64).sum().item()
-
-            mean_num = -target.numel() + ignore_count
-
-        # special mean_num is 0, return 0
-        if mean_num == 0:
-            ctx.shape = input.shape
-            ctx.mean_num = mean_num
-            ctx.dim = dim
-            ctx.input_dtype = input.dtype
-            ctx.input_device = input.device
-            if dim == 0:
-                return torch.tensor(0, device=input.device, dtype=input.dtype)
-            else:
-                return torch.tensor(
-                    float("nan"), device=input.device, dtype=input.dtype
-                )
-
-        shape = list(input.shape)
-        shape[dim] = 1
-
-        # action for target value equals ignore index, out of [0,C)
-        # 1 to delete target negetive value and set 0 to make sure scatter is OK
-        target_tmp = target
-        target = torch.where(target == ignore_index, 0, target)
-        target = torch.zeros_like(input).scatter(dim, target.view(shape), 1)
-
-        # 2 set ignore index of target value 0
-        target_tmp = target_tmp.unsqueeze(dim)
-        target_tmp = target_tmp.expand(input.shape).contiguous()
-        target = torch.where(target_tmp == ignore_index, 0, target)
-
-        M = 1
-        N = input.shape[dim]
-        for i in range(dim):
-            M *= input.shape[i]
-        inp = input.contiguous()
-        out = torch.empty_like(inp, dtype=inp.dtype)
-        K = inp.numel() // M // N
-
-        grid = lambda meta: (
-            triton.cdiv(M, meta["BLOCK_M"]),
-            K,
-        )
-        with torch.cuda.device(inp.device):
-            log_softmax_and_mul_kernel[grid](
-                out,
-                inp,
-                target,
-                mean_num,
-                M,
-                N,
-                K,
+        if weight is None:
+            weight = torch.ones(
+                [
+                    C,
+                ],
+                dtype=inp.dtype,
+                device=inp.device,
             )
-        if reduction != Reduction.NONE.value:
-            out_result = sum(out)
-        else:
-            out_result = sum_dim(out, dim=[dim])
 
-        ctx.save_for_backward(input, target)
-        ctx.dim = dim
-        ctx.mean_num = -mean_num
-        ctx.reduction = reduction
-        return out_result
+        inp = inp.contiguous()
+        tgt = target.contiguous()
+        weight = weight.contiguous()
+        out = torch.zeros(shape, dtype=torch.float32, device=inp.device)
+        grid = lambda meta: (N, triton.cdiv(D, meta["BLOCK_D"]))
+
+        if target.ndim == dim:
+            # target probabilities
+            with torch.cuda.device(inp.device):
+                celoss_probability_kernel[grid](inp, tgt, weight, out, N, C, D)
+        else:
+            # target indices
+            w_tgt = torch.zeros(shape, dtype=torch.float32, device=inp.device)
+            with torch.cuda.device(inp.device):
+                celoss_indice_kernel[grid](
+                    inp, tgt, weight, out, w_tgt, ignore_index, N, C, D
+                )
+        ctx.save_for_backward(inp, tgt, weight)
+        ctx.N = N
+        ctx.C = C
+        ctx.D = D
+        ctx.ignore_index = ignore_index
+        ctx.mean_num = 1
+        ctx.shape = shape
+
+        if reduction == 0:  # NONE
+            return out.to(inp.dtype)
+        elif reduction == 1:  # MEAN
+            if target.ndim == dim:
+                ctx.mean_num = 1 / (N * D)
+            else:
+                ctx.mean_num = 1 / sum(w_tgt).item()
+            return sum(out).to(inp.dtype) * ctx.mean_num
+        else:  # SUM
+            return sum(out).to(inp.dtype)
 
     @staticmethod
     def backward(ctx, out_grad):
         logging.debug("GEMS CrossEntropyLoss VJP")
-        mean_num = ctx.mean_num
-        dim = ctx.dim
-        if mean_num == 0:
-            input_shape = ctx.shape
-            input_dtype = ctx.input_dtype
-            input_device = ctx.input_device
-            if dim == 0:
-                return (
-                    torch.zeros(input_shape, dtype=input_dtype, device=input_device),
-                    None,
-                    None,
-                    None,
-                    None,
-                    None,
-                )
-            else:
-                return (
-                    torch.tensor(float("nan"), dtype=input_dtype, device=input_device),
-                    None,
-                    None,
-                    None,
-                    None,
-                    None,
-                )
-        input, target = ctx.saved_tensors
-        reduction = ctx.reduction
-        M = 1
-        N = input.shape[dim]
-        for i in range(dim):
-            M *= input.shape[i]
-        inp = input.contiguous()
-        out = torch.empty_like(inp, dtype=input.dtype)
-        K = inp.numel() // M // N
 
-        grid = lambda meta: (
-            triton.cdiv(M, meta["BLOCK_M"]),
-            K,
-        )
-        with torch.cuda.device(inp.device):
-            if reduction != Reduction.NONE.value:
-                softmax_and_sub_reduce_kernel[grid](
-                    out,
-                    inp,
-                    target,
-                    out_grad,
-                    mean_num,
-                    M,
-                    N,
-                    K,
-                )
-            else:
-                softmax_and_sub_kernel[grid](
-                    out,
-                    inp,
-                    target,
-                    out_grad,
-                    mean_num,
-                    M,
-                    N,
-                    K,
-                )
-        return out, None, None, None, None, None
+        inp, tgt, weight = ctx.saved_tensors
+        N = ctx.N
+        C = ctx.C
+        D = ctx.D
+        ignore_index = ctx.ignore_index
+        mean_num = ctx.mean_num
+        shape = ctx.shape
+
+        out_grad = out_grad.broadcast_to(shape).contiguous()
+
+        inp_grad = torch.zeros(inp.shape, dtype=inp.dtype, device=inp.device)
+        grid = lambda meta: (N, triton.cdiv(D, meta["BLOCK_D"]))
+        if tgt.ndim == inp.ndim:
+            celoss_probability_bwd[grid](
+                out_grad, inp, tgt, weight, inp_grad, mean_num, N, C, D
+            )
+        else:
+            celoss_indice_bwd[grid](
+                out_grad, inp, tgt, weight, inp_grad, ignore_index, mean_num, N, C, D
+            )
+        return inp_grad, None, None, None, None, None
 
 
 def cross_entropy_loss(
-    input, target, weight=None, reduction=1, ignore_index=-100, label_smoothing=0.0
+    inp, target, weight=None, reduction=1, ignore_index=-100, label_smoothing=0.0
 ):
     return CrossEntropyLoss.apply(
-        input, target, weight, reduction, ignore_index, label_smoothing
+        inp, target, weight, reduction, ignore_index, label_smoothing
     )
