+++ tests/test_reduction_ops.py
@@ -155,40 +155,75 @@ def test_accuracy_argmax(shape, dim, keepdim, dtype):
     gems_assert_equal(res_out, ref_out)
 
 
-@pytest.mark.parametrize("size_average", [None, True, False])
-@pytest.mark.parametrize("reduce", [None, True, False])
 @pytest.mark.parametrize("reduction", ["mean", "none", "sum"])
 @pytest.mark.parametrize("shape", REDUCTION_SHAPES)
 @pytest.mark.parametrize("dtype", FLOAT_DTYPES)
 @pytest.mark.parametrize("ignore_index", [1, 200, -100])
-def test_accuracy_cross_entropy_loss(
-    shape, dtype, size_average, reduce, ignore_index, reduction
-):
-    inp = torch.randn(shape, dtype=dtype, device="cuda", requires_grad=True)
+def test_accuracy_cross_entropy_loss_indices(shape, dtype, ignore_index, reduction):
     dim = 1
     up_limit = shape[dim] - 1
     target_shape = list(shape)
     del target_shape[dim]
+
+    inp = torch.randn(shape, dtype=dtype, device="cuda", requires_grad=True)
     target = torch.randint(0, up_limit, target_shape, device="cuda")
+    weight = torch.randn(shape[dim], dtype=dtype, device="cuda")
     ref_inp = to_reference(inp, True)
     ref_target = to_reference(target)
-    criterion = torch.nn.CrossEntropyLoss(
-        size_average=size_average,
-        reduce=reduce,
+    ref_weight = to_reference(weight, True)
+    ref_criterion = torch.nn.CrossEntropyLoss(
+        weight=ref_weight,
+        ignore_index=ignore_index,
+        reduction=reduction,
+    )
+    res_criterion = torch.nn.CrossEntropyLoss(
+        weight=weight,
         ignore_index=ignore_index,
         reduction=reduction,
     )
 
-    ref_out = criterion(ref_inp, ref_target)
+    ref_out = ref_criterion(ref_inp, ref_target)
     with flag_gems.use_gems():
-        res_out = criterion(inp, target)
-    gems_assert_close(res_out, ref_out, dtype)
+        res_out = res_criterion(inp, target)
+    gems_assert_close(res_out, ref_out, dtype, reduce_dim=shape[dim])
 
     out_grad = torch.randn_like(res_out)
     ref_grad = to_reference(out_grad, True)
     (ref_in_grad,) = torch.autograd.grad(ref_out, ref_inp, ref_grad)
     (res_in_grad,) = torch.autograd.grad(res_out, inp, out_grad)
-    gems_assert_close(res_in_grad, ref_in_grad, dtype)
+    gems_assert_close(res_in_grad, ref_in_grad, dtype, reduce_dim=shape[dim])
+
+
+@pytest.mark.parametrize("reduction", ["mean", "none", "sum"])
+@pytest.mark.parametrize("shape", REDUCTION_SHAPES)
+@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
+def test_accuracy_cross_entropy_loss_probabilities(shape, dtype, reduction):
+    dim = 1
+    inp = torch.randn(shape, dtype=dtype, device="cuda", requires_grad=True)
+    target = torch.randn(shape, dtype=dtype, device="cuda")
+    weight = torch.randn(shape[dim], dtype=dtype, device="cuda")
+    ref_inp = to_reference(inp, True)
+    ref_target = to_reference(target)
+    ref_weight = to_reference(weight, True)
+    ref_criterion = torch.nn.CrossEntropyLoss(
+        weight=ref_weight,
+        reduction=reduction,
+    )
+    res_criterion = torch.nn.CrossEntropyLoss(
+        weight=weight,
+        reduction=reduction,
+    )
+
+    ref_out = ref_criterion(ref_inp, ref_target)
+    with flag_gems.use_gems():
+        res_out = res_criterion(inp, target)
+    gems_assert_close(res_out, ref_out, dtype, reduce_dim=shape[dim])
+
+    out_grad = torch.randn_like(res_out)
+    ref_grad = to_reference(out_grad, True)
+    (ref_in_grad,) = torch.autograd.grad(ref_out, ref_inp, ref_grad)
+    (res_in_grad,) = torch.autograd.grad(res_out, inp, out_grad)
+    gems_assert_close(res_in_grad, ref_in_grad, dtype, reduce_dim=shape[dim])
 
 
 @pytest.mark.parametrize("shape", REDUCTION_SHAPES)
